+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 
 **Enhancements:**
 
+- Replaced `print` statements with `Log.warning` calls in `job_list.py` and `diagram.py` #2372
 - Fix characters escaped in regexes #2526
 - Fixed RO-Crate validation issues with the latest `rocrate-validator`,
   now Autosubmit crates conform to RO-Crate 1.1, process 0.5, workflow run